requirements.txt
selenium
webdriver_manager.chrome

msal==1.7.0
requests==2.25.0
pyyaml==5.3.1
python-dateutil==2.8.1